+++ src/core/sobolmatrices.h
@@ -44,8 +44,8 @@
 namespace pbrt {
 
 // Sobol Matrix Declarations
-static PBRT_CONSTEXPR int NumSobolDimensions = 1024;
-static PBRT_CONSTEXPR int SobolMatrixSize = 52;
+static PBRT_CONSTEXPR size_t NumSobolDimensions = 1024;
+static PBRT_CONSTEXPR size_t SobolMatrixSize = 52;
 extern const uint32_t SobolMatrices32[NumSobolDimensions * SobolMatrixSize];
 extern const uint64_t SobolMatrices64[NumSobolDimensions * SobolMatrixSize];
 extern const uint64_t VdCSobolMatrices[][SobolMatrixSize];
